backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
django-apscheduler>=0.6.2
pytest
pytest-django
pytest-cov
whitenoise>=6.5.0








